backend/requirements.txt
django>=5.0,<6.0
djangorestframework>=3.14
django-cors-headers>=4.3
requests>=2.31
python-dotenv>=1.0
gunicorn>=21.2
whitenoise>=6.6
